Change RemoteSegmentStoreDirectory init at given timestamp to ignore pinned timestamp setting
