+++ server/src/main/java/org/opensearch/index/store/RemoteSegmentStoreDirectory.java
@@ -196,7 +196,8 @@ public final class RemoteSegmentStoreDirectory extends FilterDirectory implement
             metadataFiles,
             Set.of(timestamp),
             MetadataFilenameUtils::getTimestamp,
-            MetadataFilenameUtils::getNodeIdByPrimaryTermAndGen
+            MetadataFilenameUtils::getNodeIdByPrimaryTermAndGen,
+            true
         );
         if (lockedMetadataFiles.isEmpty()) {
             return null;
