+++ server/src/test/java/org/opensearch/index/store/RemoteSegmentStoreDirectoryWithPinnedTimestampTests.java
@@ -8,8 +8,6 @@
 
 package org.opensearch.index.store;
 
-import org.apache.lucene.util.Version;
-import org.opensearch.common.UUIDs;
 import org.opensearch.common.blobstore.BlobMetadata;
 import org.opensearch.common.blobstore.BlobPath;
 import org.opensearch.common.blobstore.support.PlainBlobMetadata;
@@ -18,8 +16,6 @@ import org.opensearch.common.settings.Settings;
 import org.opensearch.core.action.ActionListener;
 import org.opensearch.gateway.remote.model.RemotePinnedTimestamps;
 import org.opensearch.gateway.remote.model.RemoteStorePinnedTimestampsBlobStore;
-import org.opensearch.index.remote.RemoteStoreUtils;
-import org.opensearch.index.store.remote.metadata.RemoteSegmentMetadata;
 import org.opensearch.index.translog.transfer.BlobStoreTransferService;
 import org.opensearch.indices.RemoteStoreSettings;
 import org.opensearch.node.Node;
@@ -31,7 +27,6 @@ import org.junit.Before;
 
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Supplier;
@@ -40,7 +35,6 @@ import java.util.stream.Collectors;
 import org.mockito.Mockito;
 
 import static org.opensearch.indices.RemoteStoreSettings.CLUSTER_REMOTE_STORE_PINNED_TIMESTAMP_ENABLED;
-import static org.opensearch.test.RemoteStoreTestUtils.createMetadataFileBytes;
 import static org.hamcrest.CoreMatchers.is;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.anyInt;
@@ -141,62 +135,6 @@ public class RemoteSegmentStoreDirectoryWithPinnedTimestampTests extends RemoteS
             "node-1",
             System.currentTimeMillis() - 600000
         );
-    }
-
-    public void testInitializeToSpecificTimestampNoMetadataFiles() throws IOException {
-        when(
-            remoteMetadataDirectory.listFilesByPrefixInLexicographicOrder(
-                RemoteSegmentStoreDirectory.MetadataFilenameUtils.METADATA_PREFIX,
-                Integer.MAX_VALUE
-            )
-        ).thenReturn(new ArrayList<>());
-        assertNull(remoteSegmentStoreDirectory.initializeToSpecificTimestamp(1234L));
-    }
-
-    public void testInitializeToSpecificTimestampNoMdMatchingTimestamp() throws IOException {
-        String metadataPrefix = "metadata__1__2__3__4__5__";
-        List<String> metadataFiles = new ArrayList<>();
-        metadataFiles.add(metadataPrefix + RemoteStoreUtils.invertLong(2000));
-        metadataFiles.add(metadataPrefix + RemoteStoreUtils.invertLong(3000));
-        metadataFiles.add(metadataPrefix + RemoteStoreUtils.invertLong(4000));
-
-        when(
-            remoteMetadataDirectory.listFilesByPrefixInLexicographicOrder(
-                RemoteSegmentStoreDirectory.MetadataFilenameUtils.METADATA_PREFIX,
-                Integer.MAX_VALUE
-            )
-        ).thenReturn(metadataFiles);
-        assertNull(remoteSegmentStoreDirectory.initializeToSpecificTimestamp(1234L));
-    }
-
-    public void testInitializeToSpecificTimestampMatchingMdFile() throws IOException {
-        String metadataPrefix = "metadata__1__2__3__4__5__";
-        List<String> metadataFiles = new ArrayList<>();
-        metadataFiles.add(metadataPrefix + RemoteStoreUtils.invertLong(1000));
-        metadataFiles.add(metadataPrefix + RemoteStoreUtils.invertLong(2000));
-        metadataFiles.add(metadataPrefix + RemoteStoreUtils.invertLong(3000));
-
-        Map<String, String> metadata = new HashMap<>();
-        metadata.put("_0.cfe", "_0.cfe::_0.cfe__" + UUIDs.base64UUID() + "::1234::512::" + Version.LATEST.major);
-        metadata.put("_0.cfs", "_0.cfs::_0.cfs__" + UUIDs.base64UUID() + "::2345::1024::" + Version.LATEST.major);
-
-        when(
-            remoteMetadataDirectory.listFilesByPrefixInLexicographicOrder(
-                RemoteSegmentStoreDirectory.MetadataFilenameUtils.METADATA_PREFIX,
-                Integer.MAX_VALUE
-            )
-        ).thenReturn(metadataFiles);
-        when(remoteMetadataDirectory.getBlobStream(metadataPrefix + RemoteStoreUtils.invertLong(1000))).thenReturn(
-            createMetadataFileBytes(metadata, indexShard.getLatestReplicationCheckpoint(), segmentInfos)
-        );
-
-        RemoteSegmentMetadata remoteSegmentMetadata = remoteSegmentStoreDirectory.initializeToSpecificTimestamp(1234L);
-        assertNotNull(remoteSegmentMetadata);
-        Map<String, RemoteSegmentStoreDirectory.UploadedSegmentMetadata> uploadedSegments = remoteSegmentStoreDirectory
-            .getSegmentsUploadedToRemoteStore();
-        assertEquals(2, uploadedSegments.size());
-        assertTrue(uploadedSegments.containsKey("_0.cfe"));
-        assertTrue(uploadedSegments.containsKey("_0.cfs"));
     }
 
     public void testDeleteStaleCommitsNoPinnedTimestampMdFilesLatest() throws Exception {
